Implement tag based path match as an addition to caps-based match [WIP]
This is useful if something about the stream changes that does not reflect
in the caps. One example would be some sort of extra encoding that uses
the lowest bits of PCM samples to store psychoacoustic data. These can be
played just like normal PCM, but if this extra data is present, additional
elements can be used to enhance the audio. However, notifying the presence
of this extra data in caps can be tricky, since it is essentially PCM, so
it must still be audio/x-raw data.

One approach is to add a caps field, such as "extra-encoding=true". But
the absence of such a field means that _all_ values match. So it cannot be
used for selecting a path.

One solution is to rely on tags. If specific tags are present, they
announce the presence of these special bits in the PCM samples, and an
appropriate path can be chosen.

Signed-off-by: Carlos Rafael Giani <dv@pseudoterminal.org>